Fixed compilation errors in Ubuntu 64-bit
git-svn-id: https://rapidjson.googlecode.com/svn/trunk@29 c5894555-1306-4e8d-425f-1f6f381ee07c

+++ test/unittest/valuetest.cpp
@@ -193,7 +193,7 @@ TEST(Value, Uint) {
 
 TEST(Value, Int64) {
 	// Constructor with int
-	Value x(1234LL);
+	Value x(int64_t(1234LL));
 	EXPECT_EQ(kNumberType, x.GetType());
 	EXPECT_EQ(1234, x.GetInt());
 	EXPECT_EQ(1234, x.GetUint());
@@ -214,7 +214,7 @@ TEST(Value, Int64) {
 	EXPECT_FALSE(x.IsObject());
 	EXPECT_FALSE(x.IsArray());
 
-	Value nx(-1234LL);
+	Value nx(int64_t(-1234LL));
 	EXPECT_EQ(-1234, nx.GetInt());
 	EXPECT_EQ(-1234, nx.GetInt64());
 	EXPECT_TRUE(nx.IsInt());
@@ -238,7 +238,7 @@ TEST(Value, Int64) {
 
 TEST(Value, Uint64) {
 	// Constructor with int
-	Value x(1234LL);
+	Value x(uint64_t(1234LL));
 	EXPECT_EQ(kNumberType, x.GetType());
 	EXPECT_EQ(1234, x.GetInt());
 	EXPECT_EQ(1234, x.GetUint());
